requirements.txt
# Core
python-dotenv>=1.0.0

# Gemini API (D-016) — generator, judge, and embeddings
google-genai>=0.8.0
langchain-google-genai>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10

# Vector store
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2        # local HuggingFace models (bge-*, e5-*)

# BM25 for hybrid
rank-bm25>=0.2.2

# Graph RAG
spacy>=3.7.0
networkx>=3.1

# Evaluation
ragas>=0.1.0
datasets>=2.14.0

# Data fetching
wikipedia-api>=0.6.0
arxiv>=2.0.0
requests>=2.31.0

# System metrics
psutil>=5.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# ── Elite Upgrades (v1.1) ─────────────────────────────────────────────────────
# GPU support: install PyTorch FIRST matching your CUDA version
# See: https://pytorch.org/get-started/locally/
# Then install faiss-gpu if available:
#   pip install faiss-gpu    # CUDA 11.x
# For CPU-only machines, faiss-cpu above is sufficient.

torch>=2.0.0              # GPU detection + MPS support (Apple Silicon)

# HuggingFace local models (D-013 embedding comparison benchmark)
sentence-transformers>=2.2.2
transformers>=4.35.0
huggingface-hub>=0.19.0

# tiktoken for exact token counting (optional, improves chunk sizing)
tiktoken>=0.5.0

# ── Optional: OpenAI (D-013 embedding comparison benchmark only) ──────────────
# Only needed if you want to compare OpenAI embedding models against Gemini/local.
# The core benchmark runs entirely on Gemini.
#
# openai>=1.0.0
# langchain-openai>=0.0.5
